[FP-37] Refactor: dropdownInput 컴포넌트 모듈화
